+++ __tests__/html/accessibility.keyboardHelp.clickThru.js
@@ -1,0 +1,6 @@
+/** @jest-environment ./packages/test/harness/src/host/jest/WebDriverEnvironment.js */
+
+describe('keyboard help screen', () => {
+  test('should not interfere when clicking on activities', () =>
+    runHTML('accessibility.keyboardHelp.clickThru'));
+});