+++ packages/test/page-object/src/globals/pageElements/keyboardHelpScreenCloseButton.js
@@ -1,0 +1,3 @@
+export default function keyboardHelpScreenCloseButton() {
+  return document.querySelector('button.webchat__keyboard-help__close-button[type="button"]');
+}